feat: 优化自动专精的合成部分，支持保存自己的合成方案，支持全自动专精
