+++ arknights_mower/data/agent.json
@@ -1,417 +1,1 @@
-[
-  "芬",
-  "梅",
-  "宴",
-  "砾",
-  "孑",
-  "吽",
-  "红",
-  "矩",
-  "空",
-  "黑",
-  "W",
-  "夕",
-  "林",
-  "遥",
-  "令",
-  "阿",
-  "望",
-  "黍",
-  "年",
-  "余",
-  "山",
-  "陈",
-  "锏",
-  "煌",
-  "夜刀",
-  "黑角",
-  "杜林",
-  "香草",
-  "翎羽",
-  "卡缇",
-  "斑点",
-  "炎熔",
-  "芙蓉",
-  "梓兰",
-  "夜烟",
-  "远山",
-  "协律",
-  "卡达",
-  "深靛",
-  "布丁",
-  "流星",
-  "红云",
-  "白雪",
-  "松果",
-  "酸糖",
-  "铅踝",
-  "跃跃",
-  "讯使",
-  "红豆",
-  "豆苗",
-  "冬时",
-  "杜宾",
-  "缠丸",
-  "霜叶",
-  "慕斯",
-  "刻刀",
-  "芳汀",
-  "石英",
-  "骋风",
-  "暗索",
-  "末药",
-  "清流",
-  "褐果",
-  "角峰",
-  "泡泡",
-  "露托",
-  "古米",
-  "坚雷",
-  "地灵",
-  "伊桑",
-  "阿消",
-  "维荻",
-  "云迹",
-  "微风",
-  "瑰盐",
-  "凛冬",
-  "贾维",
-  "青枳",
-  "红隼",
-  "松桐",
-  "苇草",
-  "野鬃",
-  "极境",
-  "万顷",
-  "夜半",
-  "渡桥",
-  "晓歌",
-  "谜图",
-  "寻澜",
-  "鞭刃",
-  "苍苔",
-  "医生",
-  "炎客",
-  "摩根",
-  "燧石",
-  "断崖",
-  "烈夏",
-  "铎铃",
-  "柏喙",
-  "战车",
-  "星极",
-  "铸铁",
-  "赤冬",
-  "海沫",
-  "奥达",
-  "聆音",
-  "蓝毒",
-  "白金",
-  "灰喉",
-  "四月",
-  "隐现",
-  "陨星",
-  "慑砂",
-  "截云",
-  "吉星",
-  "苦艾",
-  "雪绒",
-  "天火",
-  "惊蛰",
-  "星源",
-  "蜜蜡",
-  "寒檀",
-  "复奏",
-  "薄绿",
-  "和弦",
-  "蚀清",
-  "耶拉",
-  "洛洛",
-  "至简",
-  "折光",
-  "温米",
-  "梅尔",
-  "稀音",
-  "衡沙",
-  "赫默",
-  "亚叶",
-  "锡兰",
-  "絮雨",
-  "图耶",
-  "桑葚",
-  "蜜莓",
-  "刺玫",
-  "明椒",
-  "莎草",
-  "乌啾",
-  "风絮",
-  "临光",
-  "深律",
-  "森西",
-  "槐琥",
-  "乌有",
-  "裁度",
-  "雷蛇",
-  "深巡",
-  "可颂",
-  "拜松",
-  "火神",
-  "折桠",
-  "石棉",
-  "暮落",
-  "闪击",
-  "暴雨",
-  "灰毫",
-  "火哨",
-  "极光",
-  "洋灰",
-  "菲莱",
-  "裂响",
-  "玫拉",
-  "子月",
-  "熔泉",
-  "冰酿",
-  "雪猎",
-  "空构",
-  "崖心",
-  "雪雉",
-  "杏仁",
-  "初雪",
-  "巫恋",
-  "海霓",
-  "真理",
-  "但书",
-  "小满",
-  "掠风",
-  "海蒂",
-  "月禾",
-  "夏栎",
-  "行箸",
-  "凛视",
-  "波卜",
-  "狮蝎",
-  "绮良",
-  "罗宾",
-  "霜华",
-  "钼铅",
-  "贝娜",
-  "风丸",
-  "双月",
-  "锡人",
-  "蒂比",
-  "空弦",
-  "灰烬",
-  "鸿雪",
-  "远牙",
-  "早露",
-  "提丰",
-  "莱伊",
-  "风笛",
-  "嵯峨",
-  "忍冬",
-  "琴柳",
-  "焰尾",
-  "伺夜",
-  "异客",
-  "澄闪",
-  "黑键",
-  "维伊",
-  "妮芙",
-  "烛煌",
-  "真言",
-  "死芒",
-  "灵知",
-  "铃兰",
-  "电弧",
-  "魔王",
-  "白铁",
-  "塑心",
-  "酒神",
-  "傀影",
-  "老鲤",
-  "温蒂",
-  "水月",
-  "艾拉",
-  "闪灵",
-  "夜莺",
-  "流明",
-  "缇缇",
-  "星熊",
-  "瑕光",
-  "泥岩",
-  "斥罪",
-  "森蚺",
-  "号角",
-  "重岳",
-  "银灰",
-  "棘刺",
-  "仇白",
-  "佩佩",
-  "左乐",
-  "止颂",
-  "响石",
-  "暴行",
-  "空爆",
-  "猎蜂",
-  "杰克",
-  "夜魔",
-  "巡林者",
-  "12F",
-  "玫兰莎",
-  "泡普卡",
-  "米格鲁",
-  "克洛丝",
-  "安赛尔",
-  "格雷伊",
-  "杰西卡",
-  "安比尔",
-  "清道夫",
-  "桃金娘",
-  "断罪者",
-  "罗小黑",
-  "休谟斯",
-  "嘉维尔",
-  "苏苏洛",
-  "调香师",
-  "蛇屠箱",
-  "深海色",
-  "波登可",
-  "白面鸮",
-  "诗怀雅",
-  "芙兰卡",
-  "莱欧斯",
-  "因陀罗",
-  "达格达",
-  "幽灵鲨",
-  "布洛卡",
-  "导火索",
-  "摆渡人",
-  "羽毛笔",
-  "龙舌兰",
-  "哈蒂娅",
-  "送葬人",
-  "奥斯塔",
-  "天空盒",
-  "阿米娅",
-  "特米米",
-  "爱丽丝",
-  "戴菲恩",
-  "阿罗玛",
-  "特克诺",
-  "华法琳",
-  "录武官",
-  "诺威尔",
-  "哈洛德",
-  "卡夫卡",
-  "车尔尼",
-  "守林人",
-  "安哲拉",
-  "埃拉托",
-  "水灯心",
-  "凯瑟琳",
-  "阿兰娜",
-  "九色鹿",
-  "伯塔尼",
-  "食铁兽",
-  "见行者",
-  "若叶睦",
-  "能天使",
-  "蕾缪安",
-  "迷迭香",
-  "伊内丝",
-  "刻俄柏",
-  "逻各斯",
-  "麦哲伦",
-  "娜斯提",
-  "弑君者",
-  "多萝西",
-  "凯尔希",
-  "塞雷娅",
-  "贝洛内",
-  "赫德雷",
-  "艾丽妮",
-  "赫拉格",
-  "帕拉斯",
-  "玛恩纳",
-  "月见夜",
-  "格拉尼",
-  "斯卡蒂",
-  "安德切尔",
-  "史都华德",
-  "艾丝黛尔",
-  "罗比菈塔",
-  "德克萨斯",
-  "齐尔查克",
-  "拉普兰德",
-  "莱恩哈特",
-  "炎狱炎熔",
-  "濯尘芙蓉",
-  "普罗旺斯",
-  "格劳克斯",
-  "三角初华",
-  "八幡海铃",
-  "菲亚梅塔",
-  "维什戴尔",
-  "娜仁图亚",
-  "推进之王",
-  "缪尔赛思",
-  "可露希尔",
-  "凛御银灰",
-  "伊芙利特",
-  "莫斯提马",
-  "艾雅法拉",
-  "霍尔海雅",
-  "玛露西尔",
-  "卡涅利安",
-  "圣聆初雪",
-  "安洁莉娜",
-  "溯光星源",
-  "淬羽赫默",
-  "歌蕾蒂娅",
-  "阿斯卡纶",
-  "引星棘刺",
-  "焰影苇草",
-  "斩业星熊",
-  "丰川祥子",
-  "乌尔比安",
-  "怒潮凛冬",
-  "史尔特尔",
-  "薇薇安娜",
-  "隐德来希",
-  "司霆惊蛰",
-  "正义骑士号",
-  "历阵锐枪芬",
-  "火龙S黑角",
-  "祐天寺若麦",
-  "寒芒克洛丝",
-  "承曦格雷伊",
-  "撷英调香师",
-  "假日威龙陈",
-  "浊心斯卡蒂",
-  "麒麟R夜刀",
-  "琳琅诗怀雅",
-  "新约能天使",
-  "归溟幽灵鲨",
-  "涤火杰西卡",
-  "信仰搅拌机",
-  "百炼嘉维尔",
-  "赤刃明霄陈",
-  "耀骑士临光",
-  "圣约送葬人",
-  "荒芜拉普兰德",
-  "缄默德克萨斯",
-  "纯烬艾雅法拉",
-  "Mon3tr",
-  "THRM-EX",
-  "泰拉大陆调查团",
-  "GALLUS²",
-  "凯尔希·思衡托",
-  "维娜·维多利亚",
-  "Lancet-2",
-  "Castle-3",
-  "PhonoR-0",
-  "Friston-3",
-  "U-Official",
-  "CONFESS-47",
-  "Miss.Christine"
-]
+["芬", "梅", "宴", "砾", "孑", "吽", "红", "矩", "空", "黑", "W", "夕", "林", "遥", "令", "阿", "望", "黍", "年", "余", "山", "陈", "锏", "煌", "夜刀", "黑角", "杜林", "香草", "翎羽", "卡缇", "斑点", "炎熔", "芙蓉", "梓兰", "夜烟", "远山", "协律", "卡达", "深靛", "布丁", "流星", "红云", "白雪", "松果", "酸糖", "铅踝", "跃跃", "讯使", "红豆", "豆苗", "冬时", "杜宾", "缠丸", "霜叶", "慕斯", "刻刀", "芳汀", "石英", "骋风", "暗索", "末药", "清流", "褐果", "角峰", "泡泡", "露托", "古米", "坚雷", "地灵", "伊桑", "阿消", "维荻", "云迹", "微风", "瑰盐", "凛冬", "贾维", "青枳", "红隼", "松桐", "苇草", "野鬃", "极境", "万顷", "夜半", "渡桥", "晓歌", "谜图", "寻澜", "鞭刃", "苍苔", "医生", "炎客", "摩根", "燧石", "断崖", "烈夏", "铎铃", "柏喙", "战车", "星极", "铸铁", "赤冬", "海沫", "奥达", "聆音", "蓝毒", "白金", "灰喉", "四月", "隐现", "陨星", "慑砂", "截云", "吉星", "苦艾", "雪绒", "天火", "惊蛰", "星源", "蜜蜡", "寒檀", "复奏", "薄绿", "和弦", "蚀清", "耶拉", "洛洛", "至简", "折光", "温米", "梅尔", "稀音", "衡沙", "赫默", "亚叶", "锡兰", "絮雨", "图耶", "桑葚", "蜜莓", "刺玫", "明椒", "莎草", "乌啾", "风絮", "临光", "深律", "森西", "槐琥", "乌有", "裁度", "雷蛇", "深巡", "可颂", "拜松", "火神", "折桠", "石棉", "暮落", "闪击", "暴雨", "灰毫", "火哨", "极光", "洋灰", "菲莱", "裂响", "玫拉", "子月", "熔泉", "冰酿", "雪猎", "空构", "崖心", "雪雉", "杏仁", "初雪", "巫恋", "海霓", "真理", "但书", "小满", "掠风", "海蒂", "月禾", "夏栎", "行箸", "凛视", "波卜", "狮蝎", "绮良", "罗宾", "霜华", "钼铅", "贝娜", "风丸", "双月", "锡人", "蒂比", "空弦", "灰烬", "鸿雪", "远牙", "早露", "提丰", "莱伊", "风笛", "嵯峨", "忍冬", "琴柳", "焰尾", "伺夜", "异客", "澄闪", "黑键", "维伊", "妮芙", "烛煌", "真言", "死芒", "灵知", "铃兰", "电弧", "魔王", "白铁", "塑心", "酒神", "傀影", "老鲤", "温蒂", "水月", "艾拉", "闪灵", "夜莺", "流明", "缇缇", "星熊", "瑕光", "泥岩", "斥罪", "森蚺", "号角", "重岳", "银灰", "棘刺", "仇白", "佩佩", "左乐", "止颂", "响石", "暴行", "空爆", "猎蜂", "杰克", "夜魔", "巡林者", "12F", "玫兰莎", "泡普卡", "米格鲁", "克洛丝", "安赛尔", "格雷伊", "杰西卡", "安比尔", "清道夫", "桃金娘", "断罪者", "罗小黑", "休谟斯", "嘉维尔", "苏苏洛", "调香师", "蛇屠箱", "深海色", "波登可", "白面鸮", "诗怀雅", "芙兰卡", "莱欧斯", "因陀罗", "达格达", "幽灵鲨", "布洛卡", "导火索", "摆渡人", "羽毛笔", "龙舌兰", "哈蒂娅", "送葬人", "奥斯塔", "天空盒", "阿米娅", "特米米", "爱丽丝", "戴菲恩", "阿罗玛", "特克诺", "华法琳", "录武官", "诺威尔", "哈洛德", "卡夫卡", "车尔尼", "守林人", "安哲拉", "埃拉托", "水灯心", "凯瑟琳", "阿兰娜", "九色鹿", "伯塔尼", "食铁兽", "见行者", "若叶睦", "能天使", "蕾缪安", "迷迭香", "伊内丝", "刻俄柏", "逻各斯", "麦哲伦", "娜斯提", "弑君者", "多萝西", "凯尔希", "塞雷娅", "贝洛内", "赫德雷", "艾丽妮", "赫拉格", "帕拉斯", "玛恩纳", "月见夜", "格拉尼", "斯卡蒂", "安德切尔", "史都华德", "艾丝黛尔", "罗比菈塔", "德克萨斯", "齐尔查克", "拉普兰德", "莱恩哈特", "炎狱炎熔", "濯尘芙蓉", "普罗旺斯", "格劳克斯", "三角初华", "八幡海铃", "菲亚梅塔", "维什戴尔", "娜仁图亚", "推进之王", "缪尔赛思", "可露希尔", "凛御银灰", "伊芙利特", "莫斯提马", "艾雅法拉", "霍尔海雅", "玛露西尔", "卡涅利安", "圣聆初雪", "安洁莉娜", "溯光星源", "淬羽赫默", "歌蕾蒂娅", "阿斯卡纶", "引星棘刺", "焰影苇草", "斩业星熊", "丰川祥子", "乌尔比安", "怒潮凛冬", "史尔特尔", "薇薇安娜", "隐德来希", "司霆惊蛰", "正义骑士号", "历阵锐枪芬", "火龙S黑角", "祐天寺若麦", "寒芒克洛丝", "承曦格雷伊", "撷英调香师", "假日威龙陈", "浊心斯卡蒂", "麒麟R夜刀", "琳琅诗怀雅", "新约能天使", "归溟幽灵鲨", "涤火杰西卡", "信仰搅拌机", "百炼嘉维尔", "赤刃明霄陈", "耀骑士临光", "圣约送葬人", "荒芜拉普兰德", "缄默德克萨斯", "纯烬艾雅法拉", "Mon3tr", "THRM-EX", "泰拉大陆调查团", "GALLUS²", "凯尔希·思衡托", "维娜·维多利亚", "Lancet-2", "Castle-3", "PhonoR-0", "Friston-3", "U-Official", "CONFESS-47", "Miss.Christine"]
