+++ arknights_mower/utils/mastery_recommendation.py
@@ -264,6 +264,203 @@ def get_mastery_recommendations():
     return result
 
 
+def compute_workshop_config(t5_operator="年", book_operator="司霆惊蛰"):
+    """根据当前专精计划和仓库库存，计算合成配置（与前端自动合成配置逻辑一致）"""
+    from collections import defaultdict
+
+    from arknights_mower.data import workshop_formula
+
+    plan_path = get_path("@app/tmp/matery_plan.json")
+    planned_keys = []
+    if os.path.exists(plan_path):
+        try:
+            with open(plan_path, "r", encoding="utf-8") as f:
+                plan = json.load(f)
+            planned_keys = [k for k, v in plan.items() if v]
+        except Exception:
+            pass
+
+    skill_data_path = _find_skill_data()
+    with open(skill_data_path, "r", encoding="utf-8") as f:
+        skill_data = json.load(f)
+    items = skill_data.get("items", {})
+
+    def item_rarity(name):
+        for iid, info in items.items():
+            if info.get("name") == name:
+                return info.get("rarity", 0)
+        return 0
+
+    fodder_list = ["碳素", "碳素组", "家具零件_碳素组"]
+    fodder_items = [
+        {"item_names": [f], "children_lower_limit": 0, "self_upper_limit": 9999}
+        for f in fodder_list
+        if f in workshop_formula
+    ]
+
+    t4_names = {
+        n: e
+        for n, e in workshop_formula.items()
+        if e.get("tab") == "精英材料" and e.get("apCost") == 4.0
+    }
+    t5_names = {
+        n: e
+        for n, e in workshop_formula.items()
+        if e.get("tab") == "精英材料" and e.get("apCost") == 8.0
+    }
+
+    if not planned_keys:
+        default_t4 = [
+            {"item_names": [n], "children_lower_limit": 20, "self_upper_limit": 20}
+            for n in sorted(t4_names)
+        ]
+        default_t5 = [
+            {"item_names": [n], "children_lower_limit": 20, "self_upper_limit": 20}
+            for n in sorted(t5_names)
+        ]
+        default_book = [
+            {
+                "item_names": ["技巧概要·卷3"],
+                "children_lower_limit": 20,
+                "self_upper_limit": 20,
+            }
+        ]
+        return [
+            {"operator": "九色鹿", "enabled": True, "items": fodder_items + default_t4},
+            {"operator": t5_operator, "enabled": True, "items": default_t5},
+            {"operator": book_operator, "enabled": True, "items": default_book},
+        ]
+
+    t4_names = {
+        n: e
+        for n, e in workshop_formula.items()
+        if e.get("tab") == "精英材料" and e.get("apCost") == 4.0
+    }
+    t5_names = {
+        n: e
+        for n, e in workshop_formula.items()
+        if e.get("tab") == "精英材料" and e.get("apCost") == 8.0
+    }
+
+    if not planned_keys:
+        default_t4 = [
+            {"item_names": [n], "children_lower_limit": 20, "self_upper_limit": 20}
+            for n in sorted(t4_names)
+        ]
+        default_t5 = [
+            {"item_names": [n], "children_lower_limit": 20, "self_upper_limit": 20}
+            for n in sorted(t5_names)
+        ]
+        default_book = [
+            {
+                "item_names": ["技巧概要·卷3"],
+                "children_lower_limit": 20,
+                "self_upper_limit": 20,
+            }
+        ]
+        return [
+            {"operator": "九色鹿", "enabled": True, "items": fodder_items + default_t4},
+            {"operator": "年", "enabled": True, "items": default_t5},
+            {"operator": "司霆惊蛰", "enabled": True, "items": default_book},
+        ]
+
+    rec_result = get_mastery_recommendations()
+    operators = rec_result.get("operators", [])
+
+    plan_set = set()
+    for key in planned_keys:
+        parts = key.rsplit("_", 1)
+        if len(parts) == 2:
+            try:
+                plan_set.add((parts[0], int(parts[1])))
+            except ValueError:
+                pass
+
+    raw_demand = defaultdict(int)
+    for op in operators:
+        for rec in op.get("recommendations", []):
+            if (op["char_id"], rec["skill_index"]) not in plan_set:
+                continue
+            for mat in rec.get("chain_needed_materials", []):
+                raw_demand[mat["name"]] += mat["count"]
+
+    demand_t5_raw = {n: c for n, c in raw_demand.items() if n in t5_names}
+    demand_t4_raw = {n: c for n, c in raw_demand.items() if n in t4_names}
+    demand_t3_plus = {
+        n: c for n, c in raw_demand.items() if n not in t4_names and n not in t5_names
+    }
+
+    cultivate_path = get_path("@app/tmp/cultivate.json")
+    inventory = defaultdict(int)
+    if os.path.exists(cultivate_path):
+        with open(cultivate_path, "r", encoding="utf-8") as f:
+            cdata = json.load(f)
+        for item in cdata.get("data", {}).get("items", []):
+            cnt = int(item.get("count", 0))
+            if cnt > 0:
+                inventory[item.get("id", "")] = cnt
+
+    id_by_name = {}
+    for iid, info in items.items():
+        id_by_name[info.get("name", "")] = iid
+
+    def inv_of(name):
+        return inventory.get(id_by_name.get(name, ""), 0)
+
+    t4_indirect = defaultdict(int)
+    for t5_name, t5_demand in demand_t5_raw.items():
+        t5_missing = max(0, t5_demand - inv_of(t5_name))
+        formula = t5_names.get(t5_name, {})
+        for child in formula.get("items", []):
+            if child in t4_names:
+                t4_indirect[child] += t5_missing
+
+    t4_total = defaultdict(int)
+    for name in set(list(demand_t4_raw.keys()) + list(t4_indirect.keys())):
+        t4_total[name] = demand_t4_raw.get(name, 0) + t4_indirect.get(name, 0)
+
+    t4_items = []
+    for name, demand in sorted(t4_total.items()):
+        if demand > 0:
+            t4_items.append(
+                {
+                    "item_names": [name],
+                    "children_lower_limit": 0,
+                    "self_upper_limit": demand,
+                }
+            )
+
+    t5_items = []
+    for name, demand in sorted(demand_t5_raw.items()):
+        if demand > 0:
+            t5_items.append(
+                {
+                    "item_names": [name],
+                    "children_lower_limit": 0,
+                    "self_upper_limit": demand,
+                }
+            )
+
+    book_count = demand_t3_plus.get("技巧概要·卷3", 0)
+    book_items = (
+        [
+            {
+                "item_names": ["技巧概要·卷3"],
+                "children_lower_limit": 0,
+                "self_upper_limit": book_count,
+            }
+        ]
+        if book_count > 0
+        else []
+    )
+
+    return [
+        {"operator": "九色鹿", "enabled": True, "items": fodder_items + t4_items},
+        {"operator": "年", "enabled": True, "items": t5_items},
+        {"operator": "司霆惊蛰", "enabled": True, "items": book_items},
+    ]
+
+
 def auto_schedule_mastery_tasks():
     """仓库扫描后：检测计划内未满M3的技能，直接需求全部满足则返回待安排列表"""
     result = {"scheduled": [], "skipped": []}
